Fix dtype mismatches, edge cases, and add missing tests
## Critical Fixes

### 1. Dtype mismatch in CSA/HCA attention core (ops_csa.py, ops_hca.py)
The compression functions (csa_compress_kv, csa_compress_kv_overlapped)
return tensors in compute_dtype (fp32 for fp16 inputs). The attention
query 'q' was left in H.dtype (fp16), causing torch.einsum to raise
'RuntimeError: Expected object of scalar type Half but got scalar type
Float' for any fp16/bf16 inference. This was masked because all tests
and benchmarks used fp32/fp64.

Fix: Cast 'q' to compute_dtype before normalization in naive_csa and
naive_hca. Also cast the sliding-window branch's H_proj/C to
compute_dtype for precision consistency between the sparse/dense and
SW branches.

### 2. Dtype mismatch in HeadwiseFusedAttention (method_analysis.py)
Same issue as #1 in _csa_heads and _hca_heads. Additionally, the
forward() method concatenated KDA output (fp16) with CSA/HCA outputs
(fp32), then passed the promoted fp32 tensor to o_proj (fp16 weights),
causing 'RuntimeError: expected m1 and m2 to have the same dtype'.

Fix: Cast q to C_comp_n.dtype in _csa_heads/_hca_heads. Cast all head
outputs to x.dtype before concatenation in forward().

### 3. KDA initial_state dtype mismatch (ops_kda.py)
'S += initial_state' would raise 'RuntimeError: result type Double
can't be cast to the desired output type Float' if initial_state had
a higher precision dtype than compute_dtype. This can happen when the
caller changes dtype between calls and reuses the returned state.

Fix: Cast initial_state to compute_dtype before the in-place add, in
both naive_recurrent_kda and naive_chunk_kda.

## Edge Case Fixes

### 4. Empty timing list crashes (run_benchmark.py, run_decoding.py)
statistics.median([]) raises StatisticsError and sum([])/0 raises
ZeroDivisionError when repeats=0 or n_decode=0. Added defensive guards
that return 0.0 instead, so the JSON output stays well-formed.

### 5. make_figures.py mode selection (make_figures.py)
fig_kv_cache and fig_flops unconditionally picked 'full_accounting'
mode whenever ANY record had the accounting_mode field, silently
producing empty figures if the JSON only contained
'compressed_kv_only' records. Fixed to prefer 'full_accounting' but
fall back to 'compressed_kv_only' if that's all that exists.

### 6. Suptitle clipping in fig_ablation (make_figures.py)
tight_layout() without rect parameter could clip the suptitle at
y=1.02. Added rect=[0, 0, 1, 0.96] to reserve top margin.

### 7. n_blocks documentation (run_kv_cache.py)
Documented why max(1, T // m) is used instead of T // m for small T
(matches what a production engine actually allocates).

## New Tests

### 8. test_kda_chunk_nondivisible_T (run_correctness.py)
Verifies naive_chunk_kda matches naive_recurrent_kda when T is NOT
divisible by chunk_size. The padding code path was previously
unverified — all existing tests used T divisible by chunk_size. Tests
5 (T, chunk_size) combinations including the T=1 edge case.

### 9. test_csa_hca_fp16_dtype_consistency (run_correctness.py)
Verifies CSA and HCA run without dtype-mismatch crashes on fp16
inputs. Would have caught issue #1 before it shipped.

### 10. test_kda_initial_state_dtype_mismatch (run_correctness.py)
Verifies KDA handles initial_state with a different dtype than the
inputs. Would have caught issue #3 before it shipped.

All 54 tests pass (41 existing + 13 new).

+++ method_analysis.py
@@ -213,9 +213,12 @@ class HeadwiseFusedAttention(nn.Module):
         n_blocks = Tp // m
         C = self.csa_kv(x)
         Z = self.csa_z(x)
-        C_comp = csa_compress_kv(C, Z, self.csa_B, m)           # [B, n_blocks, c]
+        C_comp = csa_compress_kv(C, Z, self.csa_B, m)           # [B, n_blocks, c] in compute_dtype
         C_comp_n = F.normalize(C_comp, dim=-1)
-        q = F.normalize(self.csa_q(x).view(B, Tp, H, c), dim=-1)
+        # Dtype: cast q to C_comp_n's dtype so the einsum doesn't crash on
+        # fp16/bf16 inputs (C_comp is fp32 from the compression function).
+        # Mirrors the fix in ops_csa.py::naive_csa and ops_hca.py::naive_hca.
+        q = F.normalize(self.csa_q(x).view(B, Tp, H, c).to(C_comp_n.dtype), dim=-1)
         cbm = _causal_block_mask(Tp, n_blocks, m, x.device)
         scores = torch.einsum('b t h d, b n d -> b h t n', q, C_comp_n) * self.scale
         scores = scores.masked_fill(~cbm[None, None], float('-inf'))
@@ -249,7 +252,10 @@ class HeadwiseFusedAttention(nn.Module):
         Z = self.hca_z(x)
         C_comp = csa_compress_kv(C, Z, self.hca_B, m2)
         C_comp_n = F.normalize(C_comp, dim=-1)
-        q = F.normalize(self.hca_q(x).view(B, Tp, H, c), dim=-1)
+        # Dtype: cast q to C_comp_n's dtype so the einsum doesn't crash on
+        # fp16/bf16 inputs (C_comp is fp32 from the compression function).
+        # Mirrors the fix in ops_csa.py::naive_csa and ops_hca.py::naive_hca.
+        q = F.normalize(self.hca_q(x).view(B, Tp, H, c).to(C_comp_n.dtype), dim=-1)
         cbm = _causal_block_mask(Tp, n_blocks, m2, x.device)
         scores = torch.einsum('b t h d, b n d -> b h t n', q, C_comp_n) * self.scale
         scores = scores.masked_fill(~cbm[None, None], float('-inf'))
@@ -276,7 +282,18 @@ class HeadwiseFusedAttention(nn.Module):
         hca_o = self._hca_heads(h)                                  # [B, T, H_hca, c]
         # If c != hd, we need a per-head projection. For the prototype we
         # assert c == hd so concat is direct (checked in __init__).
-        all_heads = torch.cat([kda_o, csa_o, hca_o], dim=2)        # [B, T, H_total, hd]
+        #
+        # Dtype: KDA output is in x.dtype (naive_recurrent_kda casts back),
+        # but CSA/HCA outputs are in compute_dtype (fp32 for fp16 inputs)
+        # because the compression functions return fp32. Concatenating mixed
+        # dtypes promotes to fp32, which then crashes o_proj (fp16 weights)
+        # with ``RuntimeError: expected m1 and m2 to have the same dtype``.
+        # Cast all heads to x.dtype before concat so the entire forward runs
+        # in the caller's dtype. The internal fp32 computation in CSA/HCA is
+        # already done — this only affects the stored output precision.
+        all_heads = torch.cat([
+            kda_o.to(x.dtype), csa_o.to(x.dtype), hca_o.to(x.dtype)
+        ], dim=2)                                                   # [B, T, H_total, hd]
         return x + self.o_proj(all_heads.reshape(B, T, self.cfg.H_total * hd))
